refactor(server): split tap decision from mqtt transport
