+++ apps/server/src/domain/iot/index.ts
@@ -1,4 +1,5 @@
 export * from "./services/device-access-command.service";
 export * from "./services/device-command.live";
 export * from "./services/device-command.service";
+export * from "./services/device-tap-decision.service";
 export * from "./services/device-tap.service";
